requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-cors-headers==3.8.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
django-nine==0.2.5
djangorestframework==3.12.4
elasticsearch==7.14.1
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
idna==3.2
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.7
pandas==1.3.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.26.6
